refactor: Reorganize EncryptionProcessor tests into meaningful partial classes
- Consolidated 19+ fragmented test classes into 5 logical groupings
- Replaced arbitrary test splitting with functionality-based organization
- Created single EncryptionProcessorTests partial class structure:
  • CoreFunctionality.cs - End-to-end encryption/decryption, stream handling, JSON traversal
  • Validation.cs - Argument validation, settings validation, unsupported types
  • EdgeCases.cs - Depth handling, overflow scenarios, no-op operations, diagnostics
  • DataFormatEncoding.cs - ID escaping, Unicode handling, feed responses, value streams
  • Cryptography.cs - Randomized algorithms, encryption modes, key management

Benefits:
- Reduced test file fragmentation (19+ files → 6 files)
- Improved maintainability with shared utilities
- Better logical organization by functionality
- Preserved all existing test functionality
- Enhanced developer experience with meaningful test groupings
